feat: 보내요 봉투 화면 이동 (NavDeepLink 이슈)

+++ data/src/main/java/com/susu/data/data/repository/EnvelopesRepositoryImpl.kt
@@ -1,11 +1,9 @@
 package com.susu.data.data.repository
 
-import com.susu.core.model.Category
 import com.susu.core.model.FriendStatistics
 import com.susu.data.remote.api.EnvelopesService
 import com.susu.core.model.Envelope
 import com.susu.core.model.EnvelopeSearch
-import com.susu.core.model.Friend
 import com.susu.core.model.Relationship
 import com.susu.data.remote.model.request.CategoryRequest
 import com.susu.data.remote.model.request.EnvelopeRequest
@@ -17,7 +15,7 @@ class EnvelopesRepositoryImpl @Inject constructor(
     private val envelopesService: EnvelopesService,
 ) : EnvelopesRepository {
     override suspend fun getEnvelopesList(
-        friendIds: List<Int>?,
+        friendIds: List<Long>?,
         fromTotalAmounts: Int?,
         toTotalAmounts: Int?,
         page: Int?,
